requirements.txt
python-dotenv
requests
beautifulsoup4
google-generativeai
streamlit
openai
selenium
webdriver-manager
PyPDF2
